feat(server): add query handler
